Move fake_paths mock to more generic place in rtd_tests.mock.paths
